requirements.txt
Django==4.2.7
django-admin-gradient==1.0.2
django-filter==23.5
django-mdeditor==0.1.20
djangorestframework==3.14.0
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12

